Payment API Updated

+++ CollegeManagementSystem/src/main/java/com/collegemanagementsystem/Dto/TransactionHistoryDto.java
@@ -26,7 +26,7 @@ public class TransactionHistoryDto {
     private String transactionId;
     private String paymentType;
     private String registrationNo;
-    private Long totalAmount;
+    private Double totalAmount;
     private List<FeeDivisionDto> feeDivision;
 
 }
